feat(place): 관광지 목록 & 자세히 보기 View를 한 시트에 통합
- 이전은 각각 각각의 AnimatedContainer가 한 State를 공유해 페이지 이동 구현이 어려웠음
- 이를 하나의 Sheet로 통합하고 하위 3개의 view를 선택적으로 보여주게 만듬
