+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"geolocator_apple","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/geolocator_apple-2.3.7/","native_build":true,"dependencies":[]},{"name":"google_maps_flutter_ios","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/google_maps_flutter_ios-2.4.2/","native_build":true,"dependencies":[]},{"name":"url_launcher_ios","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_ios-6.2.5/","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/flutter_plugin_android_lifecycle-2.0.17/","native_build":true,"dependencies":[]},{"name":"geolocator_android","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/geolocator_android-4.5.3+1/","native_build":true,"dependencies":[]},{"name":"google_maps_flutter_android","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/google_maps_flutter_android-2.6.2/","native_build":true,"dependencies":["flutter_plugin_android_lifecycle"]},{"name":"url_launcher_android","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_android-6.3.0/","native_build":true,"dependencies":[]}],"macos":[{"name":"geolocator_apple","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/geolocator_apple-2.3.7/","native_build":true,"dependencies":[]},{"name":"url_launcher_macos","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_macos-3.1.0/","native_build":true,"dependencies":[]}],"linux":[{"name":"url_launcher_linux","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_linux-3.1.1/","native_build":true,"dependencies":[]}],"windows":[{"name":"geolocator_windows","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/geolocator_windows-0.2.3/","native_build":true,"dependencies":[]},{"name":"url_launcher_windows","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_windows-3.1.1/","native_build":true,"dependencies":[]}],"web":[{"name":"geolocator_web","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/geolocator_web-3.0.0/","dependencies":[]},{"name":"google_maps_flutter_web","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/google_maps_flutter_web-0.5.4+3/","dependencies":[]},{"name":"url_launcher_web","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_web-2.3.0/","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"geolocator","dependencies":["geolocator_android","geolocator_apple","geolocator_web","geolocator_windows"]},{"name":"geolocator_android","dependencies":[]},{"name":"geolocator_apple","dependencies":[]},{"name":"geolocator_web","dependencies":[]},{"name":"geolocator_windows","dependencies":[]},{"name":"google_maps_flutter","dependencies":["google_maps_flutter_android","google_maps_flutter_ios","google_maps_flutter_web"]},{"name":"google_maps_flutter_android","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"google_maps_flutter_ios","dependencies":[]},{"name":"google_maps_flutter_web","dependencies":[]},{"name":"url_launcher","dependencies":["url_launcher_android","url_launcher_ios","url_launcher_linux","url_launcher_macos","url_launcher_web","url_launcher_windows"]},{"name":"url_launcher_android","dependencies":[]},{"name":"url_launcher_ios","dependencies":[]},{"name":"url_launcher_linux","dependencies":[]},{"name":"url_launcher_macos","dependencies":[]},{"name":"url_launcher_web","dependencies":[]},{"name":"url_launcher_windows","dependencies":[]}],"date_created":"2024-03-21 23:06:55.176142","version":"3.19.0"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"geolocator_apple","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/geolocator_apple-2.3.7/","native_build":true,"dependencies":[]},{"name":"google_maps_flutter_ios","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/google_maps_flutter_ios-2.4.2/","native_build":true,"dependencies":[]},{"name":"url_launcher_ios","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_ios-6.2.5/","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/flutter_plugin_android_lifecycle-2.0.17/","native_build":true,"dependencies":[]},{"name":"geolocator_android","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/geolocator_android-4.5.3+1/","native_build":true,"dependencies":[]},{"name":"google_maps_flutter_android","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/google_maps_flutter_android-2.6.2/","native_build":true,"dependencies":["flutter_plugin_android_lifecycle"]},{"name":"url_launcher_android","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_android-6.3.0/","native_build":true,"dependencies":[]}],"macos":[{"name":"geolocator_apple","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/geolocator_apple-2.3.7/","native_build":true,"dependencies":[]},{"name":"url_launcher_macos","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_macos-3.1.0/","native_build":true,"dependencies":[]}],"linux":[{"name":"url_launcher_linux","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_linux-3.1.1/","native_build":true,"dependencies":[]}],"windows":[{"name":"geolocator_windows","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/geolocator_windows-0.2.3/","native_build":true,"dependencies":[]},{"name":"url_launcher_windows","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_windows-3.1.1/","native_build":true,"dependencies":[]}],"web":[{"name":"geolocator_web","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/geolocator_web-3.0.0/","dependencies":[]},{"name":"google_maps_flutter_web","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/google_maps_flutter_web-0.5.4+3/","dependencies":[]},{"name":"url_launcher_web","path":"/Users/heebeom/.pub-cache/hosted/pub.dev/url_launcher_web-2.3.0/","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"geolocator","dependencies":["geolocator_android","geolocator_apple","geolocator_web","geolocator_windows"]},{"name":"geolocator_android","dependencies":[]},{"name":"geolocator_apple","dependencies":[]},{"name":"geolocator_web","dependencies":[]},{"name":"geolocator_windows","dependencies":[]},{"name":"google_maps_flutter","dependencies":["google_maps_flutter_android","google_maps_flutter_ios","google_maps_flutter_web"]},{"name":"google_maps_flutter_android","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"google_maps_flutter_ios","dependencies":[]},{"name":"google_maps_flutter_web","dependencies":[]},{"name":"url_launcher","dependencies":["url_launcher_android","url_launcher_ios","url_launcher_linux","url_launcher_macos","url_launcher_web","url_launcher_windows"]},{"name":"url_launcher_android","dependencies":[]},{"name":"url_launcher_ios","dependencies":[]},{"name":"url_launcher_linux","dependencies":[]},{"name":"url_launcher_macos","dependencies":[]},{"name":"url_launcher_web","dependencies":[]},{"name":"url_launcher_windows","dependencies":[]}],"date_created":"2024-03-22 10:30:07.015300","version":"3.19.0"}
